delete loader/memory
